requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pydantic-settings
youtube-transcript-api
pymongo
